Began Reservation Station for functional units
This is a VERY preliminary version, posted because I've finally started to wrap my brain around a strategy. That said, at present, I think I'm trying to way too much in this version - it seems to encompass renaming and reordering, at least to some extent.  I think the reordering, at least, is unnecessary, because the instructions should be coming from the reorder buffer, and, thus, should already be reordered.  Indeed, there's a chance that I may just rework this code to be the reorder buffer and start again on the reservation station.
Nevertheless, it's a start, and any progress at all after so feeling stalled by lack of knowledge/understanding for so long feels really good.